+++ src/main/scala/fr/janalyse/primes/StreamBasedPrimesGenerator.scala
@@ -10,15 +10,6 @@ import scala.util.{ Failure, Success, Try }
 import akka.stream.scaladsl.FlowGraphImplicits
 import akka.stream.scaladsl._
 
-class NumericIterator[NUM](startFrom:NUM)(implicit numops: Integral[NUM]) extends Iterator[NUM] {
-  import numops._
-  private var num: NUM = startFrom - one
-  override def next(): NUM = {
-    num += one
-    num
-  }
-  override def hasNext(): Boolean = true
-}
 
 object StreamBasedPrimesGenerator {
   def now=System.currentTimeMillis()