+++ src/test/scala/fr/janalyse/primes/EratosthenesSieveTest.scala
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2013 David Crosson
+ * 
+ * Licensed under the GPL, Version 2.0
+ */
+
+package fr.janalyse.primes
+
+import org.junit.runner.RunWith
+import org.scalatest.FunSuite
+import org.scalatest.ShouldMatchers
+import org.scalatest.junit.JUnitRunner
+
+@RunWith(classOf[JUnitRunner])
+class EratosthenesSieveTest extends PrimesTestCommons {
+
+  test("basic tests") {
+  }
+
+}