+++ src/test/scala/fr/janalyse/primes/PerfPrimesTest.scala
@@ -18,28 +18,28 @@ class PerfPrimesTest extends PrimesTestCommons {
     val pgen = new PrimesGenerator[Long]
     import pgen._
     for (sz <- perfTestSeries)
-      howlongfor(sz, primes.drop(_).head)("lastPrime=" + _.toString)
+      howlongfor(sz)(primes.drop(_).head)("lastPrime=" + _.toString)
   }
 
   test("Performance parallel tests - Long") {
     val pgen = new PrimesGenerator[Long]
     import pgen._
     for (sz <- perfTestSeries)
-      howlongfor(sz, primesPar.drop(_).head)("lastPrime=" + _.toString)
+      howlongfor(sz)(primesPar.drop(_).head)("lastPrime=" + _.toString)
   }
 
   test("Performance classic tests - BigInt") {
     val pgen = new PrimesGenerator[BigInt]
     import pgen._
     for (sz <- perfTestSeries)
-      howlongfor(sz, primes.drop(_).head)("lastPrime=" + _.toString)
+      howlongfor(sz)(primes.drop(_).head)("lastPrime=" + _.toString)
   }
 
   test("Performance parallel tests - BigInt") {
     val pgen = new PrimesGenerator[BigInt]
     import pgen._
     for (sz <- perfTestSeries)
-      howlongfor(sz, primesPar.drop(_).head)("lastPrime=" + _.toString)
+      howlongfor(sz)(primesPar.drop(_).head)("lastPrime=" + _.toString)
   }
 
 }